organize component files
