+++ 03-react-basics/030-starting-setup/src/components/Expenses/ExpenseItem.js
@@ -1,4 +1,4 @@
-import Card from "./Card";
+import Card from "../UI/Card";
 import ExpenseDate from "./ExpenseDate";
 import "./ExpenseItem.css";
